requirements.txt
addict
gradio
ipdb
matplotlib
numpy
pandas
Pillow
PyYAML
seaborn
setuptools
termcolor
timm
torch
torchvision
transformers
yapf
